Move marketplace definitions to openhands.sdk.marketplace module
Create a new openhands.sdk.marketplace module to house marketplace-related
types (Marketplace, MarketplaceEntry, MarketplaceOwner, etc.) that were
previously embedded in openhands.sdk.plugin.types.

Changes:
- Add openhands-sdk/openhands/sdk/marketplace/ module with types.py
- Update plugin/types.py with deprecation-tracking __getattr__ re-exports
- Update plugin/__init__.py with lazy imports and deprecation warnings
- Update internal references in skills modules to use new location
- Add tests/sdk/marketplace/test_deprecation.py for deprecation tracking

The old import paths (openhands.sdk.plugin.Marketplace, etc.) still work
but emit DeprecatedWarning. Users should migrate to importing from
openhands.sdk.marketplace instead.

Deprecation timeline:
- deprecated_in: 1.16.0
- removed_in: 1.19.0
